+++ R/utils.R
@@ -15,3 +15,25 @@
     stop(paste("There is no '", layer.attr, "' vertices attribute.", sep=""))
   }
 }
+
+# considering the subgraph including only fragments connected to fragments from the same spatial unit, proportion of fragments in the 1st spatial unit (alphanumerically)
+.fragments.balance <- function(g){
+  v1 <- igraph::V(g)[igraph::V(g)$layer == unique(igraph::V(g)$layer)[1]]
+  v2 <- igraph::V(g)[igraph::V(g)$layer == unique(igraph::V(g)$layer)[2]]
+  subgraph <- igraph::subgraph_from_edges(g, igraph::E(g)[ ! v1 %--% v2 ])
+  frag.count <- table(igraph::V(subgraph)$layer)
+  round(frag.count[1] / sum(frag.count), 2)
+}
+
+
+# considering the subgraph including only fragments connected to fragments from the same spatial unit, proportion of components in the 1st spatial unit (alphanumerically)
+.components.balance <- function(g){
+  v1 <- igraph::V(g)[igraph::V(g)$layer == unique(igraph::V(g)$layer)[1]]
+  v2 <- igraph::V(g)[igraph::V(g)$layer == unique(igraph::V(g)$layer)[2]]
+  subgraph <- igraph::subgraph_from_edges(g, igraph::E(g)[ ! v1 %--% v2 ])
+  
+  compo.balance <- sapply(igraph::decompose(subgraph), 
+                          function(x) igraph::V(x)$layer[1])
+  compo.balance <- table(compo.balance)
+  round(compo.balance[1] / sum(compo.balance), 2) 
+}